requirements.txt
streamlit>=1.42.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.54
plotly>=5.24.0
scipy>=1.13.0
arch>=7.2.0
requests>=2.32.0
python-dotenv>=1.0.1
